Split analysis/recommendation/epistatic.rs into focused sub-modules (#563)
Replace the monolithic epistatic.rs (~1,500 lines) with an epistatic/ directory
containing five focused sub-modules:

- mod.rs: Public API, shared types (EpistaticPairCandidate, SynergisticCandidate,
  SourceContribution, InterferenceType, InterferencePairResult), and re-exports
- candidate_generation.rs: Pair detection, complementarity, combined improvement,
  conversion to coordinated structural candidates
- pre_screening.rs: Synergistic discovery via residual analysis (Issue #189)
- deduplication.rs: Dominant-neuron deduplication (Issue #509)
- scoring.rs: Interference detection and candidate filtering (Issue #415)

No public API changes. All existing tests pass without modification.
